lua: modify parsing of Closed Caption Data
Closed Caption (CC) Data are carried inside CDP data section. The first
value specifies the type of the channel where the CC Data has to be
extracted.
There are 2 channels. For now, only the data from the first channel is
extracted.

After a "Start" command in the CDP data, the data is extracted
(containing the "Data" command) until the first "Inactive" command.
The CC Data are stored in a Byte array where the two last bytes are
removed.